+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - `/end_encounter`: GM utility to grant pips at the end of an encounter.
 - `/damage`: GM utility to create a button allowing players to take assigned damage.
 - `/heal`: Heal yourself for a given amount, or create a button for receiving healing.
+- `/request`: GM utility to request a roll from players.
+- `/bestiary`: GM utility to view bestiary entries.
+- `/mob_attack`: GM utility to declare mob attacks.
 
 ## Bug Fixes
 - Fixed urls not being correctly parsed if they contained a `?`